requirements.txt
Flask
pandas
numpy
scikit-learn
Flask-SQLAlchemy
Flask-Login
Flask-Bcrypt
